pyproject.toml
[tool.poetry]
name = "vizing"
version = "0.1.0.9000"
description = "Constraint-based list-colouring in Python."
authors = ["Matthew Henderson <matthew.james.henderson@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-constraint = "^1.4.0"
networkx = "^3.2.1"
listcolouring = "^0.2.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
networkx = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

